+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 5.138.0 - 2018-02-16
+* Extended SemVer v2 compliance and reliability improvements - https://github.com/fsprojects/Paket/pull/3030
+* BUGFIX: Putting local folder clearing under flag in "paket clear-cache" - https://github.com/fsprojects/Paket/issues/3049
+
 #### 5.137.1 - 2018-02-14
 * BUGFIX: Allow to use different versions from different groups if they are on different frameworks - https://github.com/fsprojects/Paket/issues/3045
 * PERFORMANCE: Much faster "paket clear-cache"